+++ src/test/java/me/coley/j2h/regex/PatternHelperTest.java
@@ -1,0 +1,28 @@
+package me.coley.j2h.regex;
+
+import org.junit.Test;
+
+import static org.junit.Assert.*;
+
+public class PatternHelperTest {
+
+    private final PatternHelper gen = new PatternHelper();
+
+    @Test
+    public void testSingleRule() {
+        gen.addRule("WORD", "\\w+");
+        assertEquals("({WORD}\\w+)", getPat());
+    }
+
+    @Test
+    public void testTwoRules() {
+        gen.addRule("COLOR", "(red)|(green)|(blue)");
+        gen.addRule("NUMBR", "(one)|(two)|(three)");
+        assertEquals("({COLOR}(red)|(green)|(blue))|({NUMBR}(one)|(two)|(three))", getPat());
+    }
+
+    private String getPat() {
+        return gen.getPattern().toString();
+    }
+
+}
